@stephanlachnit 2023-11-26
Hm, doesn't seem to be faster:

Without cache:
```
real    0m16,047s
user    0m15,493s
sys     0m0,483s
```
With cache:
```
real    0m16,607s
user    0m15,640s
sys     0m0,878s
```

I'm reading the 5 different files roughly 2 times on average.